+++ far/fileview.cpp
@@ -209,7 +209,7 @@ __int64 FileViewer::VMProcess(int OpCode,void *vParam,__int64 iParam)
 
 int FileViewer::ProcessKey(int Key)
 {
-  if (RedrawTitle && ((Key & 0x00ffffff) < KEY_END_FKEY))
+  if (RedrawTitle && (((unsigned int)Key & 0x00ffffff) < KEY_END_FKEY))
     ShowConsoleTitle();
 
   if (Key!=KEY_F3 && !(Key==KEY_NUMPAD5||Key==KEY_SHIFTNUMPAD5))
